#234 refactor: ディレクトリ構成変更
