+++ src/components/Common/Pagination/index.module.css
