+++ src/components/Pages/Archive/index.tsx
@@ -1,8 +1,8 @@
 import { Article } from '@/libs/microcms';
 import Display from '@/components/Adsense/Display';
-import Pagination from '@/components/Common/Layouts/Pagination';
+import Pagination from '@/components/Common/Pagination';
 import ArticleList from '@/components/Common/ArticleList';
-import PageHeading from '@/components/Common/Layouts/PageHeading';
+import PageHeading from '@/components/Common/PageHeading';
 import Sidebar from '@/components/Common/Layouts/Sidebar';
 
 type Props = {
